New Folder Naming Panel & Bugfixes
+ New Folder Naming Panel (Like Docs)
+ Adds close new doc / new folder panel gesture (swipe down)
+ Should fix #125

This fixes #125 by means of removing underlining fuzzy search results.

This is a duct-tape fix to a problem created in:
47093c1c70034bb23b14f5ee6bf637cbd5373c21

There’s no real solution to this problem with the way documents are rendered in search results. (I guess we could but we would complicate things greatly for no good reason)

We’ll investigate, but until then we’re removing underlined search results for reasons that are too long to write in a commit message.